[#161863] Instrument quantity on reconciled statements (#3704)
# Release Notes

Reservations for reconciled order details have their `billable_minutes` set to `nil`. OSU and UMass use this attribute in their statements, which means the purchased quantity is blank. Dartmouth and NU use `PriceDisplayment#wrapped_quanity` to display the quantity on the statement. This method relies on `actual_duration_mins` and `duration_mins` instead of `billable_minutes`

This PR makes `PriceDisplayment#wrapped_statement_quanity`, which uses a calculated value of, instead of the saved value for `billable_minutes`.

This method can be used by OSU and UMass to address the blank quantity on statements.