[#162031] Refactor global price groups (#3785)
# Release Notes

This cleans up some long-standing technical cobwebs that were standing in the way of linking duration rates to price groups. 

Previously we required all price group records to have a facility ID, while also depending on a missing facility ID as a way of identifying global price groups.  This is clunky, and makes it impossible to update global price groups without skipping validations.  

With this change, price groups that are set up with `global: true` no longer require a facility ID.  This is a clearer, more standard approach, and it means the application no longer depends on invalid database records.  Plus, it paves the way for updating duration rates via price groups using `accepts_nested_attributes`.