[#162084] Improvements for stepped billing form inputs (#3808)
# Release Notes

This PR is bigger than expected, and includes a significant refactor in addition to the requirements below.
Storing min duration in its own class created a lot of unnecessary complexity.  This PR removes that and replaces it with front end code that updates hidden form inputs for duration rates.  The form does not behave well when duplicate min duration values are submitted, so I opted for additional front end validations.  There is also a uniqueness constraint on the model, but to properly handle this server-side we would need to add the concept of column index back into the mix, so I think the front end approach is better here.

- Rate is <= base rate (if Price group is the “base” group or external)
- Subsidy is <= base rate (if Price group is internal but not the “base”)
- Rate starts can't be removed if there still are Rate or Adjustment set for that step
- Show stepped billing rates on the Pricing tab for instruments (price policy index page)
- Highlight duplicate min durations in red and disable the submit form