[#162000] SciShield local storage (#3821)
# Release Notes

Some instruments require training before users are allowed to use them. The SciShield API can be used to verify a user has the proper training to use the instrument. But, if the API is down, the user will get a 500 error instead of being able to reserve the instrument.

This PR creates a local copy of a user's trainings that are checked before the API, allowing the system to be used even if the API is down.