Sync up secondary relay after any change
This keeps the logic simpler and should always be safe.
Any change results in a new relay instance, so otherwise
we'd have to do some safe navigation like:
previous_secondary_outlet = @product.relay&.secondary_outlet

Not a huge deal but it crossed the line for me of complexity vs value.